Add React-Dom dependency for storybook
I can't tell from online boards whether this is a bug or intended as
a required include for Storybook. Either way, need to be aware
moving forward that React-Dom is here, just in case that muddies
the waters for all the lerna packages and whether they require
react-dom or not.